Add GitHub Actions workflow and update build.gradle
A new workflow has been added to conduct actions when a push to the master branch is made. The workflow will check out the code, set up JDK 17, build the jar, get the version, and release. Additionally, changes have been made to the build.gradle file to support the new workflow. The version number has been bumped to '0.0.3', a new task to get the project version has been registered, and modifications have been made to the 'jar' and 'shadowJar' tasks.